+++ src/abstraction/plugin.options.ts
@@ -1,0 +1,7 @@
+import Vue from "vue";
+export interface PluginOptions  {
+    debug: boolean;
+    injections: {
+        [moduleName: string]: { [componentName: string]: Vue}
+    }
+}