+++ src/classification/index.ts
@@ -1,2 +1,3 @@
 export * from './ast.class';
 export * from './dependency.class';
+export * from './vue.di.plugin';