+++ src/function/index.ts
@@ -1,4 +1,3 @@
 export * from './function.string';
-export * from './kebab';
 export * from './pitch';
 export * from './pre.compile';