+++ src/index.ts
@@ -1,3 +1,4 @@
+export * from './abstraction';
 export * from './classification';
 export * from './function';
 export * from './globalization';