+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.<br/>
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).<br/><br/>
 
+## [0.0.5] - 2018-07-01
+* added support for default imports.
+* upgraded `vue-di-kit` to version `0.0.4`.
+* added `@types/webpack` version `4.4.4` to package.json.
+* added `VueDIPlugin` class.
+* added `PluginOptions` interface.
+
 ## [0.0.4] - 2018-06-28
 * updated webpack loader name to `vue-di-loader`, allowing webpack to properly identify the loader.
 * updated `loader.ts` to export pitch function, allowing webpack to detect the pitch function.